refactored backend added some files to gitignore and merged utils and utilities
