+++ .gitignore
@@ -1,2 +1,4 @@
 .env
 sonar-project.properties
+.idea
+.scannerwork
