--- .idea/GitLink.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="uk.co.ben_gibson.git.link.SettingsState">
-    <option name="host" value="72037fcc-cb9c-4c22-960a-ffe73fd5e229" />
-  </component>
-</project>
