--- .scannerwork/report-task.txt
@@ -1,6 +1,0 @@
-projectKey=zeitodo
-serverUrl=http://localhost:9000
-serverVersion=8.9.9.56886
-dashboardUrl=http://localhost:9000/dashboard?id=zeitodo
-ceTaskId=AYLUbsHQlgq6vTVGi9xU
-ceTaskUrl=http://localhost:9000/api/ce/task?id=AYLUbsHQlgq6vTVGi9xU
