+++ utilities/utilities_test.go
@@ -3,6 +3,7 @@ package utilities
 import (
 	"example-project/model"
 	"github.com/stretchr/testify/assert"
+	"go.mongodb.org/mongo-driver/bson/primitive"
 	"testing"
 )
 
@@ -168,4 +169,19 @@ func TestProposalTimeIntersectsProposals(t *testing.T) {
 		}
 
 	}
+}
+
+func TestGenerateToken(t *testing.T) {
+	fakeUserId := primitive.NewObjectID()
+
+	resultString := GenerateToken(fakeUserId)
+	assert.NotEmpty(t, resultString)
+}
+
+func TestValidateToken(t *testing.T) {
+	fakeToken := "fakeToken"
+
+	_, _, err := ValidateToken(fakeToken)
+
+	assert.Error(t, err)
 }
